[build, rebuild] Split common find_gemspec() out to GemspecHelpers.
